+++ LogsManager/Classes/_Extensions/Collection+Utils.swift
@@ -1,0 +1,16 @@
+//
+//  Collection+Utils.swift
+//  LogsManager
+//
+//  Created by Anton Plebanovich on 6/5/19.
+//  Copyright © 2019 Anton Plebanovich. All rights reserved.
+//
+
+import Foundation
+
+extension Collection {
+    /// Checks if collection has elements
+    var hasElements: Bool {
+        return !isEmpty
+    }
+}