+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 `LogsManager` adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [4.0.0](https://github.com/APUtils/LogsManager/releases/tag/4.0.0)
+Released on 06/05/2019.
+
+#### Changed
+- Loggers can be configured to log all components, listen to specific components or ignore some of them
+- Removed excessive flag param for error logs
+
+
 ## [3.1.0](https://github.com/APUtils/LogsManager/releases/tag/3.1.0)
 Released on 06/04/2019.